move mod11 to utils module
